rename source files to fit the convention
