+++ examples/evaluation/document-search/basic/evaluate.py
@@ -57,7 +57,7 @@ config = {
                 },
             },
             "source": {
-                "type": "ragbits.document_search.documents.hugging_face_source:HuggingFaceSource",
+                "type": "ragbits.document_search.documents.hf:HuggingFaceSource",
                 "config": {
                     "path": "micpst/hf-docs",
                     "split": "train[:5]",
