+++ packages/ragbits-core/tests/unit/vector_stores/test_in_memory.py
@@ -8,7 +8,7 @@ from ragbits.core.vector_stores.base import VectorStoreOptions
 from ragbits.core.vector_stores.in_memory import InMemoryVectorStore
 from ragbits.document_search.documents.document import DocumentMeta, DocumentType
 from ragbits.document_search.documents.element import Element
-from ragbits.document_search.documents.sources.local_file_source import LocalFileSource
+from ragbits.document_search.documents.sources.local import LocalFileSource
 
 
 class AnimalElement(Element):
