+++ packages/ragbits-document-search/src/ragbits/document_search/documents/sources/__init__.py
@@ -1,6 +1,6 @@
-from ragbits.document_search.documents.sources.gcs_source import GCSSource
-from ragbits.document_search.documents.sources.hugging_face_source import HuggingFaceSource
-from ragbits.document_search.documents.sources.local_file_source import LocalFileSource
-from ragbits.document_search.documents.sources.sources import Source
+from ragbits.document_search.documents.sources.base import Source
+from ragbits.document_search.documents.sources.gcs import GCSSource
+from ragbits.document_search.documents.sources.hf import HuggingFaceSource
+from ragbits.document_search.documents.sources.local import LocalFileSource
 
 __all__ = ["GCSSource", "HuggingFaceSource", "LocalFileSource", "Source"]
