+++ packages/ragbits-document-search/src/ragbits/document_search/documents/sources/hf.py
@@ -6,7 +6,7 @@ from typing import ClassVar
 
 from ragbits.core.utils.decorators import requires_dependencies
 from ragbits.document_search.documents.exceptions import SourceConnectionError, SourceNotFoundError
-from ragbits.document_search.documents.sources.sources import Source, get_local_storage_dir
+from ragbits.document_search.documents.sources.base import Source, get_local_storage_dir
 
 with suppress(ImportError):
     from datasets import load_dataset
