+++ packages/ragbits-document-search/tests/unit/test_source_discriminator.py
@@ -4,8 +4,8 @@ from typing import Annotated
 import pydantic
 import pytest
 
-from ragbits.document_search.documents.sources.local_file_source import LocalFileSource
-from ragbits.document_search.documents.sources.sources import Source, SourceDiscriminator
+from ragbits.document_search.documents.sources.base import Source, SourceDiscriminator
+from ragbits.document_search.documents.sources.local import LocalFileSource
 
 
 class ModelWithSource(pydantic.BaseModel):
